Defer column lazy loading after built promise
Start loading the lazy columns, when the built promise is resolved
(i.e., after the initial ranking is loaded). However, the init promise
of the view itself waits until the lazy columns are loaded.

Improvements in the loading strategy can only be achieved when
changing the ViewWrapper and tdp_core.