Patch _view_lineup.scss with older version
There is an alternative version of this file in the develop branch (with commit b9d40d04c23ef53fd2693ec8d3f93da1068a7f2d) that contains already some fixes. We copied it here now.

We do not understand yet how we got an alternative history for this file in the first place.