Increase LineUp row height to fix cell scroll bars
Scroll bars in each cell are visible using the default row height of `18px` and Bootstrap 4 default font size of `1rem`. This commit increases the row height to `20px`.

The alternative, if the row height should remain 18px, is to set/override the Boostrap 4 font size to `0.9rem`. The font size then corresponds to the one in Bootstrap 3.