Explicitly hide help link in header w/ app option
Currently the help link cannot be hidden, when the tours manager has found some tours.

This change checks if `options.showHelpLink` from the app customization is set to `false`. If so the link will be hidden.

If registered tours are available, the link with the tours modal can be shown either by setting `options.showHelpLink: true` or keep it `undefined`.

For the implementation the `options` from the constructor paramter have been used to maintain the backwards compability and show the tours modal if the `showHelpLink` is `undefined`. Using the default options would always be `false` and hide the link.