Adjust tour overlaywidth and height ...
... due to rounding errors with device pixel ratio
Usually the browser rounds the size for floating-point device pixel ratio and return a size + 1px.
In case of an overlay this additional pixel causes the display of scrollbars of the viewport.
This fix truncates the values instead of rounding it and return 1px less, depending on the device pixel ratio.

https://github.com/datavisyn/tdp_core/issues/515