Remove await `this.built` in `withoutTracking()`
Waiting for `this.built` will never resolve/terminate and show an infite loading icon.

Reason: `this.built` is created with `build()` in line 355 and it is only resolved after the `build()` call is complete.

With the changes to the selection adapter the `this.selectionAdapter.selectionChanged` now returns a valid promise. Previously, `context.add()` did not return it's promise (see https://github.com/datavisyn/tdp_core/pull/725/commits/919cfb20fb747264b425bce7a0d8a0957d3ea5ef#diff-f0f07e4a6ac30a813315b152545245d4cf26830e01b701f3e82ff7b5c22363a7R18). Now, we are waiting until all dynamic columns are added. The circle closes when we add the columns which calls `withoutTracking()` which in-turn is waiting for the resolved `this.built` promise.

As far as I can see, we always call `withoutTracking` on a built/resolved ranking view. In the other places (line 377 and 400) we have a dedicated `await this.built;` before triggering changes to dynamic columns.

In my tests adding and removing columns via score and selection adapter worked as before. I also tested reloading the page and checked the result of the session. In all cases it looks the same as before.